python-backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# LangChain & LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0

# Database & Storage
supabase>=2.9.0
psycopg2-binary>=2.9.0

# Google APIs
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.150.0

# HTTP & API Clients
httpx>=0.27.0
requests>=2.32.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.0
python-multipart>=0.0.12

# Utilities
python-dotenv>=1.0.0
tenacity>=9.0.0
python-dateutil>=2.9.0

# Logging & Monitoring
structlog>=24.0.0
pymysql>=1.1.0
PyPDF2>=3.0.0

# Google Gemini (Stage 3)
google-generativeai>=0.8.0

# Scheduling & Alignment
apscheduler>=3.10.0
thefuzz>=0.22.1
numpy<2.0.0
pypdf
cffi>=1.15.0
openai>=1.0.0